requirements.txt
boto3==1.34.65
botocore==1.34.65
certifi==2022.12.7
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
openai==1.1.1
pandas==2.2.2
pyngrok==7.0.3
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
s3transfer==0.10.1
urllib3==1.26.14
Werkzeug==3.0.1
xlrd==1.2.0
XlsxWriter==3.1.2
PyMuPDF
Pillow
pytesseract